refactor node utils
